pyproject.toml
[project]
name = "projectwhy-tts"
version = "0.1.0"
description = "Local PDF/EPUB reader with Kokoro TTS and word highlighting"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "kokoro>=0.9",
    "misaki[en]>=0.8",
    "dill>=0.3.8",
    "pypdfium2>=4",
    "doclayout-yolo>=0.0.4",
    "sounddevice>=0.5",
    "numpy>=1.26",
    "Pillow>=10",
    "PyQt6>=6.6",
    "ebooklib>=0.18",
    "beautifulsoup4>=4.12",
    "huggingface-hub>=0.20",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]

[project.scripts]
projectwhy = "projectwhy.__main__:main"

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

# Package import name is `projectwhy` under src/, not projectwhy_tts.
[tool.hatch.build.targets.wheel]
packages = ["src/projectwhy"]

[tool.ruff]
line-length = 100
